% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{adjust_wind_speed}
\alias{adjust_wind_speed}
\title{Adjust wind speed measurement to a 2-meter observation.}
\usage{
adjust_wind_speed(ws, wind_height)
}
\arguments{
\item{ws}{The wind speed in meters/second to be adjusted.}

\item{wind_height}{The height in meters of the wind speed observation.}
}
\value{
The wind speed in meters/second adjusted to a 2-meter observation.
}
\description{
Adjust wind speed measurement to a 2-meter observation.
}
\examples{
adjust_wind_speed(2.777, 10)
}
