% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_act_vapor_pressure}
\alias{calc_act_vapor_pressure}
\title{Calculate the actual vapor pressure given the saturation vapor pressure and relative humidity.}
\usage{
calc_act_vapor_pressure(svp, rh)
}
\arguments{
\item{svp}{The saturation vapor pressure in kPa. Can be calculated using \code{calc_sat_vapor_pressure}}

\item{rh}{The relative humidity (\%).}
}
\value{
The actual vapor pressure in kPa.
}
\description{
Calculate the actual vapor pressure given the saturation vapor pressure and relative humidity.
}
\examples{
calc_act_vapor_pressure(0.5, 88)
}
