% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_clear_sky_radiation}
\alias{calc_clear_sky_radiation}
\title{Calculate the clear sky radiation at a location given an elevation.}
\usage{
calc_clear_sky_radiation(elev, extra_rad)
}
\arguments{
\item{elev}{The elevation in meters.}

\item{extra_rad}{The extraterrestrial radiation in MJ m^-2 day^-1. Can be calculated using \code{calc_extraterrestrial_rad}}
}
\value{
The clear sky solar radiation in MJ m^-2 day^-1.
}
\description{
Calculate the clear sky radiation at a location given an elevation.
}
\examples{
calc_clear_sky_radiation(1000, 10)
}
