% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calc_etr_spatial}
\alias{calc_etr_spatial}
\title{Calculate spatial daily reference ET using either the Penman-Monteith
method or the Hargreaves method.}
\usage{
calc_etr_spatial(
  tmean,
  tmin = NULL,
  tmax = NULL,
  srad = NULL,
  rh = NULL,
  ws = NULL,
  elev = NULL,
  day = NULL,
  lat = NULL,
  reference = 0.23,
  z = 9,
  method = "penman"
)
}
\arguments{
\item{tmean}{A \code{terra::rast} of daily average temperature in degrees C.}

\item{tmin}{A \code{terra::rast} of  daily min temperature in degrees C.}

\item{tmax}{A \code{terra::rast} of  daily max temperature in degrees C.}

\item{srad}{A \code{terra::rast} of  daily downwelling shortwave solar radiaiton in W m^-2}

\item{rh}{A \code{terra::rast} of  daily relative humidity (\%).}

\item{ws}{A \code{terra::rast} of  daily average wind speed at 2m in M s^-1}

\item{elev}{A \code{terra::rast} of elevation in meters across the domain. If the
argument is left blank, elevation data are downloaded using the \code{elevatr}
package.}

\item{day}{The Julian day of the year ETo is being calculated for. Can either
be an integer or a date object.}

\item{lat}{A \code{terra::rast} of latitude with the same resolution, projection,
and extent as all other input rasters. If left blank, a latitude raster will
be derived from the \code{tmean} input.}

\item{reference}{The albedo of the reference surface ranging from 0 - 1.
Defaults to 0.23 for grass.}

\item{z}{The zoom level download elevation data at ranging from 1 - 14.
One if coarser resolution, 14 is finer resolution. For more information,
look at \code{?elevatr::get_elev_raster}}

\item{method}{The method to calculate ETo. Can either be "penman" or "hargreaves".}
}
\value{
A \code{terra::rast} of daily reference ET.
}
\description{
Calculate spatial daily reference ET using either the Penman-Monteith
method or the Hargreaves method.
}
\examples{
\dontrun{

# Calculate ETo across Montana for 2015-01-01
srad <- terra::rast(srad) \%>\% terra::subset(1)
tmean <- terra::rast(tmean) \%>\% terra::subset(1)
tmax <- terra::rast(tmax) \%>\% terra::subset(1)
tmin <- terra::rast(tmin) \%>\% terra::subset(1)
rh <- terra::rast(rh) \%>\% terra::subset(1)
ws <- terra::rast(ws) \%>\% terra::subset(1)

penman <- calc_etr_spatial(
  tmean = tmean, srad = srad, rh = rh, ws = ws,
  method = "penman", reference = 0.23, z = 9
)
hargreaves <- calc_etr_spatial(
  tmean = tmean, tmax = tmax, tmin = tmin, method = "hargreaves"
)
}
}
