% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calc_etr_spatial_stacked}
\alias{calc_etr_spatial_stacked}
\title{Calculate a timeseries of ETo across a set of input rasters. Assumes that all
inputs share the same resolution, extent, projection and number of layers.}
\usage{
calc_etr_spatial_stacked(
  tmean,
  tmin = NULL,
  tmax = NULL,
  srad = NULL,
  rh = NULL,
  ws = NULL,
  elev = NULL,
  days = NULL,
  reference = 0.23,
  z = 9,
  method = "penman"
)
}
\arguments{
\item{tmean}{A multilayer \code{terra::rast}}

\item{method}{}
}
\description{
Calculate a timeseries of ETo across a set of input rasters. Assumes that all
inputs share the same resolution, extent, projection and number of layers.
}
