% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_inverse_relative_distance}
\alias{calc_inverse_relative_distance}
\title{Calculate the inverse relative distance between the earth and sun for a given day.}
\usage{
calc_inverse_relative_distance(day)
}
\arguments{
\item{day}{The julian day as either an integer or date.}
}
\value{
The inverse relative distance from earth to sun.
}
\description{
Calculate the inverse relative distance between the earth and sun for a given day.
}
\examples{
calc_inverse_relative_distance(60)
}
