% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_longwave_radiation}
\alias{calc_longwave_radiation}
\title{Calculate the outgoing longwave radiation.}
\usage{
calc_longwave_radiation(temp, act_vapor_pressure, radiation_fraction)
}
\arguments{
\item{temp}{The air temperature in degrees C.}

\item{act_vapor_pressure}{The actual vapor pressure. Can be calculated with \code{calc_act_vapor_pressure}.}

\item{radiation_fraction}{The fraction of radiation relative to clear sky. Can be calculated with \code{calc_radiation_fraction}.}
}
\value{
The outgoing longwave radiation in MJ m^-2 day^-1.
}
\description{
Calculate the outgoing longwave radiation.
}
\examples{
calc_longwave_radiation(5, 1.3, 0.5)
}
