% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_pressure}
\alias{calc_pressure}
\title{Estimate the atmospheric pressure given elevation. This assumes a constant pressure of 101.3 kPa}
\usage{
calc_pressure(elev)
}
\arguments{
\item{elev}{The elevation in meters.}
}
\value{
The pressure in kPa.
}
\description{
Estimate the atmospheric pressure given elevation. This assumes a constant pressure of 101.3 kPa
}
