% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_psychrometric_constant}
\alias{calc_psychrometric_constant}
\title{Calculate the psychrometric constant given air pressure.}
\usage{
calc_psychrometric_constant(pressure)
}
\arguments{
\item{pressure}{The air pressure in kPa}
}
\value{
The psychrometric constant in kPa degC^-1
}
\description{
Calculate the psychrometric constant given air pressure.
}
