% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_radiation_fraction}
\alias{calc_radiation_fraction}
\title{Calculate fraction of radiation relative to clear sky radiation.}
\usage{
calc_radiation_fraction(radiation, clear_sky)
}
\arguments{
\item{radiation}{Incoming radiation in MJ m^-2 day^-1.}

\item{clear_sky}{The clear sky radiation in MJ m^-2 day^-1. Can be calculated with \code{calc_clear_sky_radiation}.}
}
\value{
The fraction of radiation relative to clear sky (unitless).
}
\description{
Calculate fraction of radiation relative to clear sky radiation.
}
