% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_sat_vapor_pressure}
\alias{calc_sat_vapor_pressure}
\title{Calculate the saturation vapor pressure given the daily average temperature.}
\usage{
calc_sat_vapor_pressure(temp)
}
\arguments{
\item{temp}{The temperature in degrees C.}
}
\value{
The saturation vapor pressure in kPa.
}
\description{
Calculate the saturation vapor pressure given the daily average temperature.
}
\examples{
calc_sat_vapor_pressure(5)
}
