% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_shortwave_radiation}
\alias{calc_shortwave_radiation}
\title{Calculate the outgoing shortwave radiation.}
\usage{
calc_shortwave_radiation(radiation, reference = 0.23)
}
\arguments{
\item{radiation}{The incoming radiation in MJ m^-2 day^-1.}

\item{reference}{The reference surface being used for ETo calculation. (A floating point between 0 - 1)}
}
\value{
The outgoing shortwave radiation in  MJ m^-2 day^-1.
}
\description{
Calculate the outgoing shortwave radiation.
}
\examples{
calc_shortwave_radiation(10, 0.23)
}
