% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_solar_declination}
\alias{calc_solar_declination}
\title{Given the Julian day, calculate the solar declination.}
\usage{
calc_solar_declination(day)
}
\arguments{
\item{day}{The julian day as either an integer or date.}
}
\value{
The solar declination in radians.
}
\description{
Given the Julian day, calculate the solar declination.
}
\examples{
calc_solar_declination(60)
}
