% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_sunset_hour_angle}
\alias{calc_sunset_hour_angle}
\title{Given the latitude and solar declination, calculate the sunset hour angle.}
\usage{
calc_sunset_hour_angle(lat, declination)
}
\arguments{
\item{lat}{The latitude in radians.}

\item{declination}{The solar declination in radians. Can be calculated using \code{calc_solar_declination}}
}
\value{
The solar sunset hour angle in radians
}
\description{
Given the latitude and solar declination, calculate the sunset hour angle.
}
