% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_svp_slope}
\alias{calc_svp_slope}
\title{Calculate the slope of the saturation vapor pressure curve.}
\usage{
calc_svp_slope(temp)
}
\arguments{
\item{temp}{The temperature in degrees C}
}
\value{
The slope of the saturation vapor pressure curve in kPa.
}
\description{
Calculate the slope of the saturation vapor pressure curve.
}
\examples{
calc_svp_slope(5)
}
