% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_days_from_raster}
\alias{get_days_from_raster}
\title{Create a \code{terra::rast()} of julian days from a timeseries input.}
\usage{
get_days_from_raster(r)
}
\arguments{
\item{r}{The timeseries \code{terra::rast()} used as an input.}
}
\value{
A \code{terra::rast()} that is of size \code{terra::nlyr(r)}, where each layer is a static value representing the julian day.
}
\description{
Create a \code{terra::rast()} of julian days from a timeseries input.
}
\examples{
\dontrun{
r <- terra::rast(rh)
get_days_from_raster(r)
}
}
