% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_elev_from_raster}
\alias{get_elev_from_raster}
\title{Download a raster of elevation that matches the shape, resolution and
projection of an input raster.}
\usage{
get_elev_from_raster(r, z, verbose = FALSE)
}
\arguments{
\item{r}{A \code{terra::rast} specifying the domain to download data for.}

\item{z}{The zoom level download elevation data at ranging from 1 - 14.
One if coarser resolution, 14 is finer resolution. For more information,
look at \code{?elevatr::get_elev_raster}}

\item{verbose}{Whether to display all messages when getting elevation data.
Defaults to FALSE.}
}
\value{
\code{terra::rast} of elevation in m for the input domain
}
\description{
Download a raster of elevation that matches the shape, resolution and
projection of an input raster.
}
\examples{
\dontrun{
# test
r <- terra::rast(rh)[[1]]
elev <- get_elev_from_raster(r, 9)
terra::plot(elev)
}
}
