% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_lat_from_raster}
\alias{get_lat_from_raster}
\title{Create a latitude raster from an input raster template.}
\usage{
get_lat_from_raster(r)
}
\arguments{
\item{r}{The \code{terra::rast} to generate a latitude grid for.}
}
\value{
A single layer \code{terra::rast} where each pixel's value is the latitude.
}
\description{
Create a latitude raster from an input raster template.
}
\examples{
\dontrun{
r <- terra::rast(rh)
get_lat_from_raster(r)
}
}
