% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{lat_to_radians}
\alias{lat_to_radians}
\title{Convert latitude in degrees to radians.}
\usage{
lat_to_radians(lat)
}
\arguments{
\item{lat}{The latitude in degrees to convert into radians.}
}
\value{
The latitude in radians.
}
\description{
Convert latitude in degrees to radians.
}
\examples{
lat_to_radians(45)
}
