% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmin}
\alias{tmin}
\title{Downscaled daily CMIP6 min temperature (K) projections across Montana for the
2015-01-01 to 2100-12-30 record.}
\format{
A \code{terra::rast} object with the following attributes:
\describe{
\item{CRS}{EPSG 4326}
\item{resolution}{0.25 x 0.25 degrees}
\item{time}{A date list ranging from 2015-01-01 to 2100-12-30}
}
}
\source{
\url{https://ds.nccs.nasa.gov/thredds/catalog/AMES/NEX/GDDP-CMIP6/catalog.html}
}
\usage{
tmin
}
\description{
Downscaled daily CMIP6 min temperature (K) projections across Montana for the
2015-01-01 to 2100-12-30 record.
}
\keyword{datasets}
