% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{wm2_to_mj}
\alias{wm2_to_mj}
\title{Convert radiation in watts per m^2 per day to MJ per m^2 per day.}
\usage{
wm2_to_mj(radiation)
}
\arguments{
\item{radiation}{Radiation in watts per m^2.}
}
\value{
Radiation in MJ per m^2.
}
\description{
Convert radiation in watts per m^2 per day to MJ per m^2 per day.
}
\examples{
wm2_to_mj(2)
}
