% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived.R
\name{calc_extraterrestrial_rad}
\alias{calc_extraterrestrial_rad}
\title{Calculate the theoretical extraterrestrial radiation at a given latitude.}
\usage{
calc_extraterrestrial_rad(sunset_hour, lat, declination)
}
\arguments{
\item{sunset_hour}{The sunset hour angle in radians. Can be calculated using \code{calc_sunset_hour_angle}}

\item{lat}{The latitude in radians.}

\item{declination}{The solar declination in radians. Can be calculated using \code{calc_solar_declination}}
}
\description{
Calculate the theoretical extraterrestrial radiation at a given latitude.
}
