% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_monteith.R
\name{etr_penman_monteith}
\alias{etr_penman_monteith}
\title{Calculate daily total reference ET at a given point.}
\usage{
etr_penman_monteith(lat, day, rh, temp, rad, ws, elev, reference = 0.23)
}
\arguments{
\item{lat}{Latitude of ETo calculation in degrees.}

\item{day}{The julian day ETo is being calculated on.}

\item{rh}{Relative humidity (\%).}

\item{temp}{Air temperature in Celsius.}

\item{rad}{Shortwave radiation in W m^-2.}

\item{ws}{Wind speed at 2 meters in M s^-1.}

\item{elev}{Elevation in meters.}

\item{reference}{The albedo of the reference surface (defaults to 0.23 for grass).}
}
\value{
Reference ET in mm per day.
}
\description{
Calculate daily total reference ET at a given point.
}
