% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_monteith.R
\name{etr_penman_monteith}
\alias{etr_penman_monteith}
\title{Calculate daily total reference ET at a given point. Follows the steps outlined
in Zotarelli et al (2010) \url{https://unpkg.com/node-red-contrib-evapotranspiration@1.0.1/alogorithms/evapotranspiration\%20ae45900.pdf}.}
\usage{
etr_penman_monteith(
  lat,
  days,
  rh = NULL,
  rhmin = NULL,
  rhmax = NULL,
  tmean = NULL,
  tmin = NULL,
  tmax = NULL,
  srad,
  ws,
  elev,
  reference = 0.23,
  wind_height = 2
)
}
\arguments{
\item{lat}{Latitude of ETo calculation in degrees.}

\item{rh}{Relative humidity (\%).}

\item{srad}{Shortwave radiation in W m^-2.}

\item{ws}{Wind speed at 2 meters in M s^-1.}

\item{elev}{Elevation in meters.}

\item{reference}{The albedo of the reference surface (defaults to 0.23 for grass).}

\item{wind_height}{The height of the wind observation in meters. (defaults to 2). If it is not 2 meters, the wind speed will be corrected to a 2 meter observation.}

\item{day}{The julian day ETo is being calculated on.}

\item{temp}{Air temperature in Celsius.}
}
\value{
Reference ET in mm per day.
}
\description{
Calculate daily total reference ET at a given point. Follows the steps outlined
in Zotarelli et al (2010) \url{https://unpkg.com/node-red-contrib-evapotranspiration@1.0.1/alogorithms/evapotranspiration\%20ae45900.pdf}.
}
\examples{
etr_penman_monteith(30, 150, 88, 5, 160, 5, 1000)
}
