% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_elev_from_point}
\alias{get_elev_from_point}
\title{Get the elevation in meters for a given latitude and longitude.
Latitude/longitude must be in degrees/geographic projection.}
\usage{
get_elev_from_point(lat, lon)
}
\arguments{
\item{lat}{The latitude to extract elevation for. Can either be a floating point value or a
vector of}

\item{lon}{The longitude to extract elevation for.}
}
\value{
A single value
}
\description{
Get the elevation in meters for a given latitude and longitude.
Latitude/longitude must be in degrees/geographic projection.
}
